% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bestSplit.R
\name{bestSplit}
\alias{bestSplit}
\title{bestSplit function}
\usage{
bestSplit(z)
}
\arguments{
\item{z}{data base where the last column is the dependent variable}
}
\value{
list
}
\description{
This function makes splits among columns, computes the GINI index a espefic split and select the lowest.
}
\examples{
bestSplit(iris)

}
