% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entropy.R
\name{entropy}
\alias{entropy}
\title{Entropy index}
\usage{
entropy(x)
}
\arguments{
\item{x}{Factor vector}
}
\value{
Numeric variable.
}
\description{
This function computes the entropy.
}
\examples{
entropy(iris$Species)

}
