% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/purity.R
\name{purity}
\alias{purity}
\title{purity function}
\usage{
purity(x)
}
\arguments{
\item{x}{Factor vector}
}
\value{
string
}
\description{
This function defines if a vector is pure
}
\examples{
purity(iris$Species)

}
