% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sspf.R
\name{sspf}
\alias{sspf}
\title{sspf function}
\usage{
sspf(y, x)
}
\arguments{
\item{y}{Factor}

\item{x}{vector}
}
\value{
data.frame
}
\description{
This function makes splits, computes the GINI index a espefic split and select the lowest.
}
\examples{
sspf(iris$Species,iris$Sepal.Width)

}
