% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treeCl.R
\name{treeCl}
\alias{treeCl}
\title{treeCl function}
\usage{
treeCl(formula, data, deep = NULL)
}
\arguments{
\item{formula}{Formula to define the dependent and independent varibles}

\item{data}{data base}

\item{deep}{Numeric value to indicate the deep of the tree}
}
\value{
tree object
}
\description{
This function uses several functions to build a tree based on data
}
\examples{

treeCl(Species~Petal.Width+Petal.Length+Sepal.Width,data = iris, deep = 3)

myTree <- treeCl(Species~Petal.Width+Petal.Length+Sepal.Width,data = iris, deep = 3)

plot(myTree)

}
