% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treeClass.R
\docType{package}
\name{treeClass}
\alias{treeClass}
\alias{treeClass-package}
\title{treeClass Package}
\description{
treeClass is a package for data classification based on GINI index.
}
\author{
Ernesto Rubio \email{ernestorub@gmail.com}

Aleksi Tavkhelidze \email{aleksi.tavkhelidze@uni-goettingen.de}
}
