% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bestSplit.R
\name{bestSplit}
\alias{bestSplit}
\title{bestSplit function}
\usage{
bestSplit(z)
}
\arguments{
\item{z}{database, where the last column is the dependent variable}
}
\value{
list with a data.frame called base, the best (i.e. lowest) Gini Index, the value at which the split was made and the variable corresponding to the best split
}
\description{
This function makes splits among columns, computes the Gini index for a specific split and selects the column with the lowest GINI Index.
}
\examples{
bestSplit(iris)

}
