% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bestSplit.R
\name{bestSplit}
\alias{bestSplit}
\title{bestSplit function}
\usage{
bestSplit(z)
}
\arguments{
\item{z}{database where the last column is the dependent variable}
}
\value{
list with a data.frame called base, the best GINI index, the split value and the variable with the best split
}
\description{
This function makes splits among columns, computes the GINI index a espefic split and select the column with the lowest GINI.
}
\examples{
bestSplit(iris)

}
