% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sspf.R
\name{sspf}
\alias{sspf}
\title{sspf function}
\usage{
sspf(y, x)
}
\arguments{
\item{y}{Factor vector}

\item{x}{vector: it can be numeric, a character or a factor}
}
\value{
data.frame
}
\description{
This function splits the independent variable and computes the weighted Gini index by selecting the lowest one
}
\examples{
sspf(iris$Species,iris$Sepal.Width)

}
