% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treeCl.R
\name{treeCl}
\alias{treeCl}
\title{treeCl function}
\usage{
treeCl(formula, data, depth)
}
\arguments{
\item{formula}{Formula to define the dependent and independent varibles}

\item{data}{database}

\item{depth}{Numeric value to indicate the depth of the tree}
}
\value{
tree object
}
\description{
This function uses several functions to build a classification tree based on data
}
\examples{

treeCl(Species~Petal.Width+Petal.Length+Sepal.Width,data = iris, depth = 3)

myTree <- treeCl(Species~Petal.Width+Petal.Length+Sepal.Width,data = iris, depth = 3)

plot(myTree)

}
